Add Registry abstraction to mediate Storage and Packages
In the process of trying to test integration I ran into a bunch
of issues that I just needed to knock out while I had the code in
front of me. In particular, I believe I was able to make significant
progress in decoupling the backend storage from the cran server
logic. This commit adds a new class called Registry which does things
like parse DESCRIPTION files and create PACKAGES files.

There's a ton of stuff in here that's getting snuck in on a testing
PR but I want to get this out and I don't have much time to pick things
apart right now.